Separate styles from tsx file
